# backend/openapi/credence_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Credence AI",
    "version": "1.0.0"
  },
  "paths": {
    "/wallet/analyze": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Analyze Wallet",
        "operationId": "analyze_wallet_wallet_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/credit/score": {
      "post": {
        "tags": [
          "Credit"
        ],
        "summary": "Score Wallet",
        "operationId": "score_wallet_credit_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/protocol/create-loan": {
      "post": {
        "tags": [
          "Loan Protocol"
        ],
        "summary": "Create On Chain Loan",
        "operationId": "create_on_chain_loan_protocol_create_loan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/repayment": {
      "post": {
        "tags": [
          "Repayment"
        ],
        "summary": "Repay Loan",
        "description": "Submits a request to repay an active loan on-chain.\nReturns transaction success details or structured failure.",
        "operationId": "repay_loan_repayment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reputation/{wallet}": {
      "get": {
        "tags": [
          "Reputation"
        ],
        "summary": "Get Reputation",
        "description": "Retrieves the authoritative reputation profile, trust score, rating,\nand history of versioned profiles for a borrower address.",
        "operationId": "get_reputation_reputation__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/integrations/{protocol}/{wallet}": {
      "get": {
        "tags": [
          "Protocol Integrations"
        ],
        "summary": "Get Protocol Integration",
        "description": "Consolidates the borrower's reputation state and executes the selected\nprotocol adapter to return a standardized, time-bound integration contract.",
        "operationId": "get_protocol_integration_integrations__protocol___wallet__get",
        "parameters": [
          {
            "name": "protocol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Protocol"
            }
          },
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/developer/api-key": {
      "post": {
        "tags": [
          "Developer Platform"
        ],
        "summary": "Create Api Key",
        "description": "Generates a secure live or test API key.",
        "operationId": "create_api_key_developer_api_key_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/developer/api-key/revoke": {
      "post": {
        "tags": [
          "Developer Platform"
        ],
        "summary": "Revoke Api Key",
        "description": "Revokes / deactivates a specific API key.",
        "operationId": "revoke_api_key_developer_api_key_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/developer/webhooks/register": {
      "post": {
        "tags": [
          "Developer Platform"
        ],
        "summary": "Register Webhook",
        "description": "Registers a new webhook subscription endpoint.",
        "operationId": "register_webhook_developer_webhooks_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/developer/webhooks/unregister": {
      "post": {
        "tags": [
          "Developer Platform"
        ],
        "summary": "Unregister Webhook",
        "description": "Unregisters a webhook subscription.",
        "operationId": "unregister_webhook_developer_webhooks_unregister_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/profiles/{wallet}": {
      "get": {
        "tags": [
          "Credit Profiles"
        ],
        "summary": "Get Credit Profile Catalog",
        "description": "Consolidates credit and reputation metrics into a unified developer profile.",
        "operationId": "get_credit_profile_catalog_profiles__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyRequest": {
        "properties": {
          "is_live": {
            "type": "boolean",
            "title": "Is Live",
            "default": false
          }
        },
        "type": "object",
        "title": "ApiKeyRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepaymentRequest": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "loan_id": {
            "type": "string",
            "title": "Loan Id"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "loan_id"
        ],
        "title": "RepaymentRequest"
      },
      "RevokeKeyRequest": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "RevokeKeyRequest"
      },
...
      "UnregisterWebhookRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "UnregisterWebhookRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalletRequest": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          }
        },
        "type": "object",
        "required": [
          "wallet"
        ],
        "title": "WalletRequest"
      },
      "WebhookRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "events": {
            "items": {},
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "url",
          "events"
        ],
        "title": "WebhookRequest"
      },
...
    }
  }
}